+++ src/land_pools.py
@@ -24,12 +24,31 @@ def update_land_pools_for_account(account, land_pools):
     return land_pools
 
 
+def update_land_resource(land_resources):
+    do_update = False
+    if not land_resources.empty:
+        date = datetime.today().strftime('%Y-%m-%d')
+        if land_resources.loc[land_resources.date == date].empty:
+            do_update = True
+        else:
+            logging.warning("Already pulled land resource data for this day, no need to run this too often")
+    else:
+        do_update = True
+
+    if do_update:
+        df1 = land_util.get_land_resources_info()
+        land_resources = pd.concat([land_resources, df1], ignore_index=True)
+    return land_resources
+
+
 def update_pools():
-    progress_util.update_daily_msg("Start update land pools")
+    progress_util.update_daily_msg("Start update land")
 
     for account in store_util.get_account_names():
         progress_util.update_daily_msg("...update land pools for: " + str(account))
         store.land_pools = update_land_pools_for_account(account, store.land_pools.copy())
+
+    store.land_resources = update_land_resource(store.land_resources.copy())
 
     store.land.fillna(0, inplace=True)
     store_util.save_stores()