+++ CHANGES.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Update season SPS graph (add validator rewards)
 - Update hive blog generation with validator node rewards
+- Update land pools (DEC-WOOD, DEC-STONE, DEC-IRON)
 
 ## 0.24.4